Add sync verification and repair after exceptions.
Having determined that exceptions during the record update process can cause inserts that aren't committed in the database, we now watch for such exceptions and do re-verification of the sync state when they occur.  We also monitor at startup and every 20 minutes to make sure they two key tables - contacts and fundraising - are still in sync.