Fix ordering of updates in update_changed_records.
The original ordering collected all the records before uploading any, but some of the records depend on the earlier uploads.

Also added regular update checks to the housekeeping loop so that changes made manually to the data are reliably pushed to Airtable.

Also added a way of invoking the incremental update via stv.py