Remove overlap between 2021 and 2022 donations.
As per spec, we have been crediting donations since 11/1/2021 as "2022 donations" for the purposes of classifying their donors as contacts and including them in the Donations table.  But we were also including them in the 2021 donation summaries and totals, causing them to be double-counted.  The decision has been made to eliminate the double counting and to consider them exclusively as 2022 donations.