Fix obscure issue with annual recurring donations.
Annual recurring donations can happen to people who have never been marked as donors before (because their last donation was before November of last year).  When we notice this, we have to force-recompute their status.

This fix feels a bit tortured, but it will work even if there are other bugs that miss marking donors.  Since we've had quite a few of these, it's worth using the elephant gun.

Probably we should also handle annual recurring donations by marking the donor as a donor in the first place, even if we haven't seen a recurrence yet this year.  But we're not worrying about that fix for the 2 annual recurring donors we have.

This gets a patch release bump.