Add coordinated events to integration database and Airtable.
This permits coordinated events to appear on the STV calendar feed, but does not provide any attendance data for those events.

There is also a "Coordinated?*" checkbox in the Events table to make clear which events are coordinated.

This gets a minor version bump (and dependency update) for new functionality.