Fix missing assignments bug.
The "2020 Assignments*" field in the Contacts table was empty because it was using a lookup from Contacts to Historical Contacts (which we should never have made).  This fix updates the integration to populate that field in Contacts directly.

This is a patch release for the bug fix.  Dependencies are now at latest.