Fix logic around preferring refcodes to page owners for attribution.
We weren't checking to see if the refcode was associated with a person before preferring it to the page owner.  Now, if the refcode is not associated, we fall back to using the page owner.  This allows the use of unknown refcodes on supporter pages not to interfere with the attribution to the suppoter.